Major cleaning; replaced sed with new --override option in genrun.
genrun.py --override={"key": val} overrides the value in the
passed json config file, removing the need to mutate that file
directly with `sed`.  This exposes the (key,value) pair directly to
swift, which is nice.

Removed all case-specific references in the code: RTI, viscosity, etc.
Everything is specified outside of the main loop.