Replaced iterate with foreach, changed the way nek handles filenames.
At the suggestion of @hategan, replaced iterate with foreach to
workaround bug where iterate serializes loop body executions.

foreach has the added benefit of facilitating contiuation jobs,
so the starting iteration index can now be passed in as j0.

Also changed the nek code to accept a session (e.g. debug-0, debug-1,
etc...) and a series (e.g. debug).  The session is used to name
the rea, and map file while the series is used to name the outputs
and checkpoints.  This avoids the large checkpoint copies, at the cost
of breaking backwards compatability with Nek5000.